Update_food_delivery_time_prediction.py
The vehicle type is defined as a dummy variable. 
Dummy variables that fit the vehicle type were determined with the scatter plot. 
After that function created to convert the data type as an int which entried data as a str for vehicle type. 
Finally this variable added to model as a int.